moved SingleFileAutoAnnotateEndToEndTestHelper to test utils
